Rename equipment module to production planning
