--- lib/modules/equipment/equipment_screen.dart
@@ -1,18 +1,0 @@
-import 'package:flutter/material.dart';
-
-class EquipmentScreen extends StatelessWidget {
-  const EquipmentScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Оборудование')),
-      body: const Center(
-        child: Text(
-          'Здесь будет управление оборудованием',
-          style: TextStyle(fontSize: 18),
-        ),
-      ),
-    );
-  }
-}
